requirements.txt
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.3
autopep8==1.5
billiard==3.3.0.23
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
Django==2.2.12
django-cors-headers==3.2.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
drf-extensions==0.6.0
entrypoints==0.3
flake8==3.7.9
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
mccabe==0.6.1
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
progress==1.5
psycopg2-binary==2.8.5
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==1.7.1
pyparsing==2.4.6
python-dotenv==0.13.0
pytoml==0.1.21
pytz==2019.3
requests==2.23.0
retrying==1.3.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
webencodings==0.5.1
whitenoise==5.0.1
